fix: selenium test
