+++ app/modules/dataset/tests/test_selenium_rate_datasets.py
@@ -1,0 +1,52 @@
+# Generated by Selenium IDE
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.chrome.service import Service
+from selenium.webdriver.chrome.options import Options
+
+
+class TestRatedatasets():
+  def setup_method(self, method):
+    # Set up options for Chromium
+    chrome_options = Options()
+    chrome_options.binary_location = '/usr/bin/chromium-browser'  # Update this path as needed
+
+    # Create a service object for the ChromeDriver
+    service = Service('/usr/bin/chromedriver')  # Ensure this is the correct path to your ChromeDriver
+
+    # Initialize the WebDriver with the Chromium options
+    self.driver = webdriver.Chrome(service=service, options=chrome_options)
+    self.vars = {}
+
+  def teardown_method(self, method):
+    self.driver.quit()
+
+  def test_ratedatasets(self):
+    self.driver.get("http://127.0.0.1:5000/")
+    self.driver.set_window_size(1050, 731)
+    self.driver.find_element(By.LINK_TEXT, "Explore").click()
+    self.driver.find_element(By.ID, "search").click()
+    self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(1)").click()
+    self.driver.find_element(By.ID, "email").click()
+    self.driver.find_element(By.ID, "email").send_keys("user1@example.com")
+    self.driver.find_element(By.ID, "password").click()
+    self.driver.find_element(By.ID, "password").send_keys("1234")
+    self.driver.find_element(By.ID, "submit").click()
+    self.driver.find_element(By.LINK_TEXT, "View dataset").click()
+    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-12 > span:nth-child(5)").click()
+    self.driver.find_element(By.ID, "rate-button-12").click()
+
+    self.driver.find_element(By.CSS_SELECTOR, ".text-dark").click()
+    self.driver.find_element(By.CSS_SELECTOR, ".dropdown-menu > .dropdown-item:nth-child(2)").click()
+    
+    self.driver.get("http://127.0.0.1:5000/")
+        
+    self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(1)").click()
+    self.driver.find_element(By.ID, "email").click()
+    self.driver.find_element(By.ID, "email").send_keys("user2@example.com")
+    self.driver.find_element(By.ID, "password").click()
+    self.driver.find_element(By.ID, "password").send_keys("1234")
+    self.driver.find_element(By.ID, "submit").click()
+    self.driver.find_element(By.LINK_TEXT, "View dataset").click()
+    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-12 > span:nth-child(1)").click()
+    self.driver.find_element(By.ID, "rate-button-12").click()
