--- app/modules/dataset/tests/test_selenium_rating.py
@@ -1,81 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-class TestRating():
-  def setup_method(self, method):
-    self.driver = webdriver.Chrome()
-    self.vars = {}
-
-  def teardown_method(self, method):
-    self.driver.quit()
-
-  def test_rating(self):
-    self.driver.get("http://127.0.0.1:5000/")
-    self.driver.set_window_size(1129, 715)
-    self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(2)").click()
-    self.driver.find_element(By.ID, "name").click()
-    self.driver.find_element(By.ID, "name").send_keys("Giorno")
-    self.driver.find_element(By.ID, "surname").click()
-    self.driver.find_element(By.ID, "surname").send_keys("Giovanna")
-    self.driver.find_element(By.ID, "email").click()
-    self.driver.find_element(By.ID, "email").send_keys("correo@email.com")
-    self.driver.find_element(By.ID, "password").click()
-    self.driver.find_element(By.ID, "password").send_keys("contraseña")
-    self.driver.find_element(By.ID, "submit").click()
-    self.driver.find_element(By.LINK_TEXT, "View dataset").click()
-    self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(1)").click()
-    self.driver.find_element(By.ID, "email").click()
-    self.driver.find_element(By.ID, "email").send_keys("correo@email.com")
-    self.driver.find_element(By.ID, "password").click()
-    self.driver.find_element(By.ID, "password").send_keys("contraseña")
-    self.driver.find_element(By.ID, "submit").click()
-    self.driver.find_element(By.LINK_TEXT, "View dataset").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-8 > span:nth-child(4)").click()
-    self.driver.find_element(By.ID, "rate-button-8").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-    self.driver.find_element(By.CSS_SELECTOR, ".sidebar-item:nth-child(3) .align-middle:nth-child(2)").click()
-    self.driver.find_element(By.ID, "search").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-7 > span:nth-child(2)").click()
-    self.driver.find_element(By.ID, "rate-button-7").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-    self.driver.find_element(By.LINK_TEXT, "Explore").click()
-    self.driver.find_element(By.ID, "search").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-6 > span:nth-child(4)").click()
-    self.driver.find_element(By.ID, "rate-button-6").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-    self.driver.find_element(By.CSS_SELECTOR, ".text-dark").click()
-    self.driver.find_element(By.CSS_SELECTOR, ".dropdown-menu > .dropdown-item:nth-child(2)").click()
-    self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(2)").click()
-    self.driver.find_element(By.ID, "name").click()
-    self.driver.find_element(By.ID, "name").send_keys("Elvis")
-    self.driver.find_element(By.ID, "surname").click()
-    self.driver.find_element(By.ID, "surname").send_keys("Presley")
-    self.driver.find_element(By.ID, "email").click()
-    self.driver.find_element(By.ID, "email").send_keys("elvis@presley.sing")
-    self.driver.find_element(By.ID, "password").click()
-    self.driver.find_element(By.ID, "password").send_keys("contraseña")
-    self.driver.find_element(By.ID, "submit").click()
-    self.driver.find_element(By.CSS_SELECTOR, ".card:nth-child(4) .btn:nth-child(1)").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-5 > span:nth-child(1)").click()
-    self.driver.find_element(By.ID, "rate-button-5").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-    self.driver.find_element(By.LINK_TEXT, "Explore").click()
-    self.driver.find_element(By.ID, "search").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-6 > span:nth-child(1)").click()
-    self.driver.find_element(By.ID, "rate-button-6").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-    self.driver.find_element(By.CSS_SELECTOR, ".sidebar-item:nth-child(3) .align-middle:nth-child(2)").click()
-    self.driver.find_element(By.ID, "search").click()
-    self.driver.find_element(By.CSS_SELECTOR, "#star-rating-7 > span:nth-child(5)").click()
-    self.driver.find_element(By.ID, "rate-button-7").click()
-    assert self.driver.switch_to.alert.text == "Rating submitted successfully!"
-
